@@ -6,7 +6,7 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ^3.5.0
+  sdk: ^3.7.2
 
 dependencies:
   flutter:
@@ -14,10 +14,10 @@ dependencies:
   get: ^4.6.6
   flutter_easyloading: ^3.0.5
   google_fonts: ^6.2.1
-  intl: ^0.19.0
+  intl: ^0.20.2
   flutter_masked_text2: ^0.9.1
-  pinch_zoom_release_unzoom: ^1.0.3
-  flutter_slidable: ^3.1.1
+  pinch_zoom_release_unzoom: ^2.0.0
+  flutter_slidable: ^4.0.0
   flutter_credit_card: ^4.0.1
   carousel_slider: ^5.0.0
   
@@ -27,9 +27,9 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  flutter_lints: ^4.0.0
+  flutter_lints: ^5.0.0
   flutter_native_splash: ^2.4.1
-  flutter_launcher_icons: ^0.13.1
+  flutter_launcher_icons: ^0.14.3
 
 
 flutter_native_splash:
