@@ -35,7 +35,7 @@ class _CategoryProductCardState extends State<CategoryProductCard> {
               offset: const Offset(0, 0),
               spreadRadius: 1,
               blurRadius: 1,
-              color: Colors.black.withOpacity(0.1),
+              color: Colors.black.withValues(alpha: 0.1),
             )
           ],
         ),
