@@ -3,10 +3,8 @@ import UIKit
 import XCTest
 
 class RunnerTests: XCTestCase {
-
-  func testExample() {
-    // If you add code to the Runner application, consider adding tests here.
-    // See https://developer.apple.com/documentation/xctest for more information about using XCTest.
-  }
-
+    func testExample() {
+        // If you add code to the Runner application, consider adding tests here.
+        // See https://developer.apple.com/documentation/xctest for more information about using XCTest.
+    }
 }
